Fixed Webpack issues, began html structure
Ran into some issues with the changes from Webpack 4 to 5. The node
command to run the dev server is now "webpack serve ..." instead of
"webpack-dev-server". I also ran into problems with the file structure
and webpack not reading past the src folder for the entry point,
resulting in making the index.js entry point a direct child of src.

I began structuring out the basic template for the application. It is a
rough cut of where everything will go before the tune-ups along the way.