Modified WebPack config and dependencies/Began JS
I had problems using the output.library through webpack in dev mode.
Currently, it works in production but I am unable to access functions in
development. I also found optimize-css-assets-webpack-plugin to be
deprecated and swapped in the css-minimizer-webpack-plugin instead.

The JS has begun, and I am starting with the overlay that pops up when
an input is requested to add a new trip.