Integrated APIs and Found a Bug w/ dev-server/HMR
Integrated functionality with Geonames, Weatherbit, and Pixabay APIs.

-Geonames takes in the city name and responds with a list of places. The
App will chose the first option.

-Weatherbit will get historical weather data for the longitude and
latitude provided from Geonames because it doesn't serve predictions too
far out. Since the Weatherbit account is on the free version, it needs
to call the API 7 times for a weeks worth of historical data.

-Pixabay takes in the city and responds with an image for that location.

As for the WebPack bug, there is an issue with wp5, webpack-dev-server
and hot module replacement. Essentially it causes the library to empty,
not allowing access to imported functions.
The issue is found here:
https://github.com/webpack/webpack-dev-server/issues/2484
and offers a fix setting config.devServer.injectClient: false.
This ends up disabling hot reload it seems but I can access the library
and reload the page manually to see changes.

I also added miscellaneous style changes as well as a countdown in days
for the trips.