Add/Remove location functions, added weather icons
The app now has the ability to add multiple locations to each trip.
Every location also receives a 'remove location' button that asks
for confirmation before removing the location. If it is the last one,
the entire trip is removed.

A file loader was added in order to add weather icons to the weather
section of each location. Icons were used from iconfinder.com with free
use licenses.

I have begun styling the app, putting on the final touches.