Added EsLint, Jest, and Service Workers, and style
Setup the linter with the eslint recommended options. I created a couple
tests for the app. 1 for the server, and the other to test removing a
trip. I installed the updated version of the Workbox-Wepback-plugin and
setup service workers for the site to load offline. Finally, I added
style to the site and wrote the README.